--- src/main/java/org/novak/java/dto/WorkspaceCreateRequestDTO.java
@@ -1,23 +1,0 @@
-package org.novak.java.dto;
-
-import jakarta.validation.constraints.DecimalMax;
-import jakarta.validation.constraints.DecimalMin;
-import jakarta.validation.constraints.NotNull;
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-import org.novak.java.model.workspace.WorkspaceType;
-
-@Data
-@NoArgsConstructor
-@AllArgsConstructor
-public class WorkspaceCreateRequestDTO {
-
-    @NotNull(message = "Price must not be null")
-    @DecimalMin(value = "0.01", message = "Price must be greater than 0")
-    @DecimalMax(value = "1499.99", message = "Price must be less than 1500")
-    private Double price;
-
-    @NotNull(message = "Workspace type must not be null")
-    private WorkspaceType workspaceType;
-}
